frontend(tidal): extract TidalDiscoverShelves reusable renderer
